+++ src/Sephrasto/CharakterVorteileWrapper.py
@@ -30,7 +30,7 @@ class CharakterVorteileWrapper(QtCore.QObject):
         
         self.uiVor.splitter.adjustSize()
         width = self.uiVor.splitter.size().width()
-        self.uiVor.splitter.setSizes([width*0.6, width*0.4])
+        self.uiVor.splitter.setSizes([int(width*0.6), int(width*0.4)])
 
         if len(Wolke.Char.vorteile) > 0:
             self.currentVort = Wolke.Char.vorteile[0]
