+++ src/Sephrasto/RuestungPicker.py
@@ -42,7 +42,7 @@ class RuestungPicker(object):
 
         self.ui.splitter.adjustSize()
         width = self.ui.splitter.size().width()
-        self.ui.splitter.setSizes([width*0.6, width*0.4])
+        self.ui.splitter.setSizes([int(width*0.6), int(width*0.4)])
         
         self.ruestungsTypen = Wolke.DB.einstellungen["Rüstungen: Typen"].toTextList()
